feat(p4-sp3): DocumentConfigPage — tabbed shell for 9 doc types
Rewrites DocumentConfigPage as a 9-tab UI (numbering + 8 doc-type tabs).
Extracts original P2-SP2 numbering/prefix content into NumberingConfigTab.
Tab keys are snake_case (numbering, deposit_receipt, receipt, ...).
URL ?tab=<key> for deep-link navigation.
